+++ molecool/tests/test_molecool.py
@@ -7,6 +7,7 @@ import molecool
 import pytest
 import sys
 
+
 def test_molecool_imported():
     """Sample test, will always pass so long as import statement worked"""
     assert "molecool" in sys.modules
